# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WhatsApp Web Panel API",
    "version": "1.0.0",
    "description": "Cookie tabanli oturum (express-session) + CSRF korumali API. /api altindaki mutating (POST/PUT/DELETE) endpoint'ler icin CSRF header gonderilmelidir."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "whatsapp.sid"
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      },
      "xsrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-XSRF-TOKEN"
      }
    },
    "parameters": {
      "AccountIdHeader": {
        "name": "X-Account-Id",
        "in": "header",
        "required": false,
        "schema": { "type": "string" },
        "description": "Multi-account icin aktif hesabi secmekte kullanilir."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string" },
          "requestId": { "type": "string" }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "required": ["success"],
        "properties": {
          "success": { "type": "boolean" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/accounts": {
      "get": {
        "tags": ["Accounts"],
        "summary": "Accounts list (admin)",
        "security": [{ "sessionCookie": [] }],
        "responses": { "200": { "description": "Accounts", "content": { "application/json": { "schema": { "type": "object" } } } } }
      },
      "post": {
        "tags": ["Accounts"],
        "summary": "Account create (admin)",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["name"], "properties": { "name": { "type": "string" } } } } }
        },
        "responses": { "200": { "description": "Created", "content": { "application/json": { "schema": { "type": "object" } } } } }
      }
    },
    "/api/accounts/select": {
      "post": {
        "tags": ["Accounts"],
        "summary": "Active account select (admin)",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["accountId"], "properties": { "accountId": { "type": "string" } } } } }
        },
        "responses": { "200": { "description": "Selected", "content": { "application/json": { "schema": { "type": "object" } } } } }
      }
    },
    "/api/chats": {
      "get": {
        "tags": ["Chats"],
        "summary": "Chats list",
        "security": [{ "sessionCookie": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "tag", "in": "query", "required": false, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Chats", "content": { "application/json": { "schema": { "type": "array", "items": { "type": "object" } } } } } }
      }
    },
    "/api/chats/search": {
      "get": {
        "tags": ["Chats"],
        "summary": "Chats search",
        "security": [{ "sessionCookie": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "q", "in": "query", "required": false, "schema": { "type": "string" } },
          { "name": "tag", "in": "query", "required": false, "schema": { "type": "string" } },
          { "name": "note", "in": "query", "required": false, "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer" } },
          { "name": "offset", "in": "query", "required": false, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Results", "content": { "application/json": { "schema": { "type": "array", "items": { "type": "object" } } } } } }
      }
    },
    "/api/chats/{chatId}/messages": {
      "get": {
        "tags": ["Chats"],
        "summary": "Chat messages + tags + notes",
        "security": [{ "sessionCookie": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "chatId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer" } },
          { "name": "offset", "in": "query", "required": false, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Messages", "content": { "application/json": { "schema": { "type": "object" } } } } }
      }
    },
...
    "/api/chats/{chatId}/tags": {
      "get": {
        "tags": ["Chats"],
        "summary": "Chat tags",
        "security": [{ "sessionCookie": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "chatId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Tags", "content": { "application/json": { "schema": { "type": "array", "items": { "type": "object" } } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "post": {
        "tags": ["Chats"],
        "summary": "Chat tag ekle",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "chatId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["tag_id"],
                "properties": { "tag_id": { "type": "integer" } }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/chats/{chatId}/tags/{tagId}": {
      "delete": {
        "tags": ["Chats"],
        "summary": "Chat tag kaldir",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "chatId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "tagId", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/chats/{chatId}/notes": {
      "get": {
        "tags": ["Chats"],
        "summary": "Chat notes",
        "security": [{ "sessionCookie": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "chatId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Notes", "content": { "application/json": { "schema": { "type": "array", "items": { "type": "object" } } } } }
        }
      },
      "post": {
        "tags": ["Chats"],
        "summary": "Chat note ekle",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "chatId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["content"],
                "properties": { "content": { "type": "string" } }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/chats/{chatId}/notes/{noteId}": {
      "put": {
        "tags": ["Chats"],
        "summary": "Chat note guncelle",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "chatId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "noteId", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["content"],
                "properties": { "content": { "type": "string" } }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } }
        }
      },
      "delete": {
        "tags": ["Chats"],
        "summary": "Chat note sil",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "chatId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "noteId", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } }
        }
      }
    },
...
    "/api/auto-replies": {
      "get": {
        "tags": ["Scheduled"],
        "summary": "Auto replies list",
        "security": [{ "sessionCookie": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AccountIdHeader" }],
        "responses": { "200": { "description": "Auto replies", "content": { "application/json": { "schema": { "type": "array", "items": { "type": "object" } } } } } }
      },
      "post": {
        "tags": ["Scheduled"],
        "summary": "Auto reply create (admin/manager)",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AccountIdHeader" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["trigger_word"], "properties": { "trigger_word": { "type": "string" }, "response": { "type": "string" }, "template_id": { "type": ["integer", "null"] }, "match_type": { "type": "string" }, "is_active": { "type": "boolean" }, "required_tag_id": { "type": ["integer", "null"] }, "exclude_tag_id": { "type": ["integer", "null"] } } } } } },
        "responses": { "200": { "description": "Created", "content": { "application/json": { "schema": { "type": "object" } } } } }
      }
    },
    "/api/auto-replies/{id}": {
      "put": {
        "tags": ["Scheduled"],
        "summary": "Auto reply update (admin/manager)",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } },
        "responses": { "200": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } } }
      },
      "delete": {
        "tags": ["Scheduled"],
        "summary": "Auto reply delete (admin/manager)",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } } }
      }
    },
    "/api/auto-replies/{id}/toggle": {
      "post": {
        "tags": ["Scheduled"],
        "summary": "Auto reply toggle (admin/manager)",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Success", "content": { "application/json": { "schema": { "type": "object" } } } } }
      }
    },
...
  }
}